Fixed erroneous Clojure/MQL namespace specifications.
It appears as though a bad string replace messed up some of the Strings
representing MQL namespaces. The compiler didn't catch this as the
behavior doesn't manifest itself until a query is registered.

This fixes Netflix/mantis#27 but the prudent thing would be to pull the
MQL wrapper into the mql-jvm package to reduce the possibility of this
error occuring again.